requirements.txt
tensorflow==2.12.0
pandas 
python-dotenv
dvc
pillow
mlflow==2.2.2
notebook
numpy
matplotlib
seaborn
streamlit
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
scipy
fastapi
gdown
uvicorn
pydantic
python-multipart
-e .